requirements.txt
asgiref==3.8.1
black==24.10.0
click==8.1.7
colorama==0.4.6
Django==5.1.3
django-environ==0.11.2
django-filter==24.3
django-graphql-jwt==0.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
flake8==7.1.1
graphene==3.4.1
graphene-django==3.2.2
graphql-core==3.2.5
graphql-relay==3.2.0
gunicorn==23.0.0
iniconfig==2.0.0
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.3.6
pluggy==1.5.0
promise==2.3
psycopg==3.2.3
psycopg-binary==3.2.3
psycopg2-binary==2.9.10
pycodestyle==2.12.1
pyflakes==3.2.0
PyJWT==2.9.0
pytest==8.3.3
pytest-django==4.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
six==1.16.0
sqlparse==0.5.1
strawberry-graphql==0.247.2
strawberry-graphql-django==0.49.1
text-unidecode==1.3
typing_extensions==4.12.2
tzdata==2024.2
whitenoise==6.8.2
